[수정] 데이터 한곳에서 우선 모아서 쓸 수 있도록 구조 개선
